# docs/generated/openapi.json
{
  "components": {
    "schemas": {
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScenarioNextActionRequest": {
        "additionalProperties": false,
        "properties": {
          "checkpoint_id": {
            "title": "Checkpoint Id",
            "type": "string"
          }
        },
        "required": [
          "checkpoint_id"
        ],
        "title": "ScenarioNextActionRequest",
        "type": "object"
      },
      "ScenarioSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "allowed_next_actions": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Next Actions",
            "type": "array"
          },
          "current_checkpoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Checkpoint Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "result_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Artifact Id"
          },
          "scenario_session_id": {
            "title": "Scenario Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "scenario_session_id",
          "job_id",
          "status"
        ],
        "title": "ScenarioSessionResponse",
        "type": "object"
      },
      "ScenarioStartRequest": {
        "additionalProperties": false,
        "properties": {
          "frontend_id": {
            "title": "Frontend Id",
            "type": "string"
          },
          "guest_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Id"
          },
          "input": {
            "title": "Input"
          },
          "source_frontend_instance_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Frontend Instance Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "frontend_id",
          "input"
        ],
        "title": "ScenarioStartRequest",
        "type": "object"
      },
      "ScenarioStartResponse": {
        "additionalProperties": false,
        "properties": {
          "allowed_next_actions": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Next Actions",
            "type": "array"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "result_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Artifact Id"
          },
          "scenario_session_id": {
            "title": "Scenario Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "scenario_session_id",
          "job_id",
          "status"
        ],
        "title": "ScenarioStartResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AnytoolAI Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/products/{product_id}/scenarios/{scenario_id}/start": {
      "post": {
        "operationId": "start_scenario_v1_products__product_id__scenarios__scenario_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "allowed_next_actions": [],
                  "job_id": "job_123",
                  "scenario_session_id": "scenario_session_123",
                  "status": "started"
                },
                "schema": {
                  "$ref": "#/components/schemas/ScenarioStartResponse"
                }
              }
            },
            "description": "Stable queue-and-return response for CE polling."
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "guest_not_found": {
                    "summary": "Supplied guest identity is unknown.",
                    "value": {
                      "error": {
                        "code": "guest_identity_not_found",
                        "message": "Guest identity not found.",
                        "request_id": "req_123"
                      }
                    }
                  },
                  "scenario_not_found": {
                    "summary": "Scenario is unknown for the product.",
                    "value": {
                      "error": {
                        "code": "scenario_not_found",
                        "message": "Scenario not found.",
                        "request_id": "req_123"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe response when the scenario is unknown for the product or the supplied guest identity is unknown."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "idempotency_key_conflict",
                    "message": "Idempotency-Key was already used with a different request.",
                    "request_id": "req_123"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Idempotency-Key was reused with a request whose scope or input does not match the original request. No new scenario session is created and quota is not consumed; retry with a new Idempotency-Key or the original request."
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "guest_required": {
                    "summary": "Quota-protected product requires guest_id.",
                    "value": {
                      "error": {
                        "code": "guest_identity_required",
                        "message": "Guest identity is required for this product.",
                        "request_id": "req_123"
                      }
                    }
                  },
                  "idempotency_key_invalid": {
                    "summary": "Idempotency-Key exceeds the maximum length.",
                    "value": {
                      "error": {
                        "code": "idempotency_key_invalid",
                        "message": "Idempotency-Key must be at most 256 characters.",
                        "request_id": "req_123"
                      }
                    }
                  },
                  "invalid_input": {
                    "summary": "Scenario input is not a JSON object.",
                    "value": {
                      "error": {
                        "code": "scenario_input_invalid",
                        "message": "Scenario input must be a JSON object.",
                        "request_id": "req_123"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe validation response for unsupported frontend, invalid input, missing guest identity, or an oversized Idempotency-Key."
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "quota_exhausted",
                    "message": "Guest quota exhausted.",
                    "request_id": "req_123"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Quota exhausted. The backend rejects the scenario start before creating any scenario session or linked job."
          }
        },
        "summary": "Create a scenario session and queue workflow execution",
        "tags": [
          "scenario-runtime"
        ]
      }
    },
...
    "/v1/scenario-sessions/{scenario_session_id}": {
      "get": {
        "operationId": "get_scenario_session_v1_scenario_sessions__scenario_session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_session_id",
            "required": true,
            "schema": {
              "title": "Scenario Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "allowed_next_actions": [
                    "copy_result",
                    "create_handoff"
                  ],
                  "current_checkpoint_id": "result_ready",
                  "job_id": "job_123",
                  "result_artifact_id": "artifact_123",
                  "scenario_session_id": "scenario_session_123",
                  "status": "completed"
                },
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSessionResponse"
                }
              }
            },
            "description": "Frontend-safe session snapshot for polling."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "scenario_session_not_found",
                    "message": "Scenario session not found.",
                    "request_id": "req_123"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe response when the scenario session is unknown."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get frontend-safe scenario session state",
        "tags": [
          "scenario-runtime"
        ]
      }
    },
    "/v1/scenario-sessions/{scenario_session_id}/next-actions/{next_action_id}": {
      "post": {
        "operationId": "post_next_action_v1_scenario_sessions__scenario_session_id__next_actions__next_action_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_session_id",
            "required": true,
            "schema": {
              "title": "Scenario Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "next_action_id",
            "required": true,
            "schema": {
              "title": "Next Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioNextActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "allowed_next_actions": [
                    "copy_result",
                    "create_handoff"
                  ],
                  "current_checkpoint_id": "result_ready",
                  "job_id": "job_123",
                  "result_artifact_id": "artifact_123",
                  "scenario_session_id": "scenario_session_123",
                  "status": "completed"
                },
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSessionResponse"
                }
              }
            },
            "description": "Validated next action with the current safe session snapshot."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "scenario_session_not_found",
                    "message": "Scenario session not found.",
                    "request_id": "req_123"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe response when the scenario session is unknown."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "scenario_checkpoint_conflict",
                    "message": "Scenario checkpoint no longer matches the requested action.",
                    "request_id": "req_123"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe response for stale checkpoint or disallowed action."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate and record a frontend next-action click",
        "tags": [
          "scenario-runtime"
        ]
      }
    }
  }
}